Fix update rate calculation crash

+++ VRCOSC.Game/Modules/Module.cs
@@ -47,8 +47,8 @@ public abstract class Module
     protected Player Player = null!;
     protected ChatBox ChatBox = null!;
 
-    public const int vrc_osc_update_rate = 20;
-    public static readonly int vrc_osc_delta_update = (1 / vrc_osc_update_rate) * 1000;
+    public const float vrc_osc_update_rate = 20;
+    public static readonly int vrc_osc_delta_update = (int)((1f / vrc_osc_update_rate) * 1000f);
 
     public void Initialise(GameHost host, Storage storage, OscClient oscClient, ChatBox chatBox)
     {
